requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
kombu==5.3.5
mypy==1.9.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
prompt-toolkit==3.0.43
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
